requirements.txt
# -e git+ssh://git@github.com/norraell/hackathon-big-berlin-hack.git@470aa4953ad8dea0ad7671984ca2bca06ae7ad0b#egg=ai_claims_intake
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
attrs==26.1.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==47.0.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.1
frozenlist==1.8.0
google-ai-generativelanguage==0.6.10
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-cloud-speech==2.28.0
google-generativeai==0.8.3
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.2
httptools==0.7.1
idna==3.13
Mako==1.3.11
MarkupSafe==3.0.3
multidict==6.7.1
numpy==2.4.4
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
PyJWT==2.12.1
pyparsing==3.3.2
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
requests==2.33.1
SQLAlchemy==2.0.49
starlette==1.0.0
tqdm==4.67.3
twilio==9.10.5
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
yarl==1.23.0
